pushing delete function

+++ server.py
@@ -40,12 +40,13 @@ def image_return():
 
 @app.route('/<address>.png', methods=['GET'])
 def unique_image_return(address):
-    with open('{}.png'.format(address), 'rb') as f:
-        data = f.read()
-    resp = Response(data, mimetype='image/png')
-    os.remove('{}.wtml'.format(address))
-    os.remove('{}.png'.format(address))
-    return resp
+    if '{}.png'.format(address).exists():
+        with open('{}.png'.format(address), 'rb') as f:
+            data = f.read()
+        resp = Response(data, mimetype='image/png')
+        os.remove('{}.wtml'.format(address))
+        os.remove('{}.png'.format(address))
+        return resp
     # try:
     #     return send_file('{}.png'.format(address), mimetype='image/png', cache_timeout=1)
     # except:
